refactor(#114): game 패키지 AttendanceService/Vote → GameParticipation으로 리네이밍
- game/AttendanceVote → game/GameParticipation
- game/AttendanceService → game/GameParticipationService
- infrastructure/service/game/AttendanceServiceImpl → GameParticipationServiceImpl
- Bounded Context 명확화로 import 모호성 해소
- DB 스키마(테이블명 attendance_votes)는 변경 없음
- 모든 참조 경로 업데이트 (Repository, Controller, Mapper, Tests)
