+++ nextup-infrastructure/src/main/kotlin/com/nextup/infrastructure/repository/AttendanceVoteRepository.kt
@@ -1,42 +1,42 @@
 package com.nextup.infrastructure.repository
 
 import com.nextup.core.domain.game.AttendanceStatus
-import com.nextup.core.domain.game.AttendanceVote
+import com.nextup.core.domain.game.GameParticipation
 import com.nextup.core.port.repository.AttendanceVoteRepositoryPort
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Query
 
 /**
- * AttendanceVote Repository
+ * GameParticipation Repository
  * JpaRepository 상속 + Port 인터페이스 구현
  */
 interface AttendanceVoteRepository :
-    JpaRepository<AttendanceVote, Long>,
+    JpaRepository<GameParticipation, Long>,
     AttendanceVoteRepositoryPort {
-    override fun findByIdOrNull(id: Long): AttendanceVote? = findById(id).orElse(null)
+    override fun findByIdOrNull(id: Long): GameParticipation? = findById(id).orElse(null)
 
-    @Query("SELECT av FROM AttendanceVote av WHERE av.game.id = :gameId")
-    override fun findByGameId(gameId: Long): List<AttendanceVote>
+    @Query("SELECT av FROM GameParticipation av WHERE av.game.id = :gameId")
+    override fun findByGameId(gameId: Long): List<GameParticipation>
 
-    @Query("SELECT av FROM AttendanceVote av WHERE av.game.id = :gameId AND av.member.id = :memberId")
+    @Query("SELECT av FROM GameParticipation av WHERE av.game.id = :gameId AND av.member.id = :memberId")
     override fun findByGameIdAndMemberId(
         gameId: Long,
         memberId: Long,
-    ): AttendanceVote?
+    ): GameParticipation?
 
-    @Query("SELECT av FROM AttendanceVote av WHERE av.game.id = :gameId AND av.status = :status")
+    @Query("SELECT av FROM GameParticipation av WHERE av.game.id = :gameId AND av.status = :status")
     override fun findByGameIdAndStatus(
         gameId: Long,
         status: AttendanceStatus,
-    ): List<AttendanceVote>
+    ): List<GameParticipation>
 
-    @Query("SELECT av FROM AttendanceVote av WHERE av.game.id = :gameId AND av.respondedAt IS NULL")
-    override fun findNonVotersByGameId(gameId: Long): List<AttendanceVote>
+    @Query("SELECT av FROM GameParticipation av WHERE av.game.id = :gameId AND av.respondedAt IS NULL")
+    override fun findNonVotersByGameId(gameId: Long): List<GameParticipation>
 
-    @Query("SELECT COUNT(av) FROM AttendanceVote av WHERE av.game.id = :gameId")
+    @Query("SELECT COUNT(av) FROM GameParticipation av WHERE av.game.id = :gameId")
     override fun countByGameId(gameId: Long): Long
 
-    @Query("SELECT COUNT(av) FROM AttendanceVote av WHERE av.game.id = :gameId AND av.status = :status")
+    @Query("SELECT COUNT(av) FROM GameParticipation av WHERE av.game.id = :gameId AND av.status = :status")
     override fun countByGameIdAndStatus(
         gameId: Long,
         status: AttendanceStatus,
